refactor: Moved all neural network components into a single folder.
