+++ frontend/src/components/neural_network/Neuron.tsx
