+++ frontend/src/components/neural_network/NeuronLayer.tsx
